Add filter to only record slow db queries

+++ src/helpers.php
@@ -156,7 +156,7 @@ function prepare_http_response( $response ) {
 
 function prepare_wpdb_query( array $query_array ) : array {
 	$query = isset( $query_array[0] ) ? $query_array[0] : '';
-	$duration = isset( $query_array[1] ) ? (int) ( $query_array[1] * 1000 ) : 0;
+	$duration = isset( $query_array[1] ) ? ( $query_array[1] * 1000 ) : 0;
 
 	return [ $query, $duration ];
 }
